ref: Move logic out of App.
